requirements.txt
easygui==0.98.3
numpy==1.23.5
Pillow==9.4.0
rembg==2.0.31
streamlit==1.20.0
